+++ test/test-typings.ts
@@ -89,6 +89,7 @@ let isGoalTracked: Record<string, boolean> = vwoInstance.track(
 );
 
 let isPostDimensionPushed = vwoInstance.push('tagKey', 'tagValue', 'userId');
+vwoInstance.push({'a': 'a'}, 'userId');
 
 ///flush batch events manually