+++ test/test-utils/data/settingsFile-11.js
@@ -1,0 +1,135 @@
+module.exports = {
+  groups: {},
+  campaignGroups: {},
+  campaigns: [
+    {
+      id: 160,
+      segments: {},
+      status: 'RUNNING',
+      percentTraffic: 100,
+      goals: [
+        {
+          identifier: 'track',
+          id: 284,
+          type: 'CUSTOM_GOAL'
+        },
+        {
+          identifier: 'track2',
+          id: 285,
+          type: 'CUSTOM_GOAL'
+        },
+        {
+          identifier: 'track3',
+          id: 286,
+          type: 'REVENUE_TRACKING',
+          revenueProp: 'abcd'
+        }
+      ],
+      isForcedVariationEnabled: false,
+      key: 'track1',
+      name: 'track1',
+      variations: [
+        {
+          id: 1,
+          name: 'Control',
+          changes: {},
+          weight: 50
+        },
+        {
+          id: 2,
+          name: 'Variation-1',
+          changes: {},
+          weight: 50
+        }
+      ],
+      type: 'FEATURE_TEST'
+    },
+    {
+      id: 161,
+      segments: {},
+      status: 'RUNNING',
+      percentTraffic: 100,
+      goals: [
+        {
+          identifier: 'track',
+          id: 1,
+          type: 'REVENUE_TRACKING'
+        },
+        {
+          identifier: 'track3',
+          id: 286,
+          type: 'REVENUE_TRACKING',
+          revenueProp: 'abcdefg'
+        },
+        {
+          identifier: 'track2',
+          id: 287,
+          type: 'CUSTOM_GOAL'
+        }
+      ],
+      isForcedVariationEnabled: false,
+      key: 'track2',
+      name: 'track2',
+      variations: [
+        {
+          id: 1,
+          name: 'Control',
+          changes: {},
+          weight: 50
+        },
+        {
+          id: 2,
+          name: 'Variation-1',
+          changes: {},
+          weight: 50
+        }
+      ],
+      type: 'VISUAL_AB'
+    },
+    {
+      id: 162,
+      segments: {},
+      status: 'RUNNING',
+      percentTraffic: 100,
+      goals: [
+        {
+          identifier: 'track',
+          id: 1,
+          type: 'REVENUE_TRACKING'
+        },
+        {
+          identifier: 'track3',
+          id: 286,
+          type: 'REVENUE_TRACKING',
+          revenueProp: 'abcd'
+        },
+        {
+          identifier: 'track2',
+          id: 287,
+          type: 'CUSTOM_GOAL'
+        }
+      ],
+      isForcedVariationEnabled: false,
+      key: 'track3',
+      name: 'track3',
+      variations: [
+        {
+          id: 1,
+          name: 'Control',
+          changes: {},
+          weight: 50
+        },
+        {
+          id: 2,
+          name: 'Variation-1',
+          changes: {},
+          weight: 50
+        }
+      ],
+      type: 'VISUAL_AB'
+    }
+  ],
+  accountId: 1,
+  isEventArchEnabled: true,
+  version: 1
+};