+++ test/test-utils/data/settingsFiles.js
@@ -8,6 +8,7 @@ const settingsFile7 = require('./settingsFile-7');
 const settingsFile8 = require('./settingsFile-8');
 const settingsFile9 = require('./settingsFile-9');
 const settingsFile10 = require('./settingsFile-10');
+const settingsFile11 = require('./settingsFile-11');
 const settingsFileMeg = require('./settingsFile-meg');
 
 module.exports = {
@@ -21,5 +22,6 @@ module.exports = {
   settingsFile8,
   settingsFile9,
   settingsFile10,
+  settingsFile11,
   settingsFileMeg
 };